Move proxy struct into proxy.go
